Refactoring of Shop Dashboard
