+++ kor-react/package.json
@@ -4,6 +4,7 @@
   "private": true,
   "dependencies": {
     "@auth0/auth0-react": "^2.4.0",
+    "@posthog/react": "^1.2.3",
     "@testing-library/dom": "^10.4.1",
     "@testing-library/jest-dom": "^6.8.0",
     "@testing-library/react": "^16.3.0",
@@ -15,7 +16,7 @@
     "@types/react-helmet": "^6.1.11",
     "axios": "^1.11.0",
     "lucide-react": "^0.541.0",
-    "posthog-js": "^1.268.9",
+    "posthog-js": "^1.276.0",
     "react": "^19.1.1",
     "react-dom": "^19.1.1",
     "react-helmet": "^6.1.0",
