+++ kor-react/src/components/shop/modules/FamilyPlanModules.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import { DashboardSectionProps } from '../types';
+
+const FamilyPlanModules: React.FC<DashboardSectionProps> = ({ shopUser, planFeatures }) => {
+  return (
+  <div></div> //Blank - Update with custom modules later
+  )
+};
+
+export default FamilyPlanModules;
