+++ kor-react/src/components/shop/modules/index.ts
@@ -1,0 +1,2 @@
+export { default as FamilyPlanModules } from './FamilyPlanModules';
+export { default as PlanSpecificModules } from './PlanSpecificModules';
