+++ kor-react/src/components/shop/sections/index.ts
@@ -1,0 +1,6 @@
+export { default as WelcomeHeader } from './WelcomeHeader';
+export { default as PlanFeaturesBanner } from './PlanFeaturesBanner';
+export { default as GettingStartedSection } from './GettingStartedSection';
+export { default as QRCodeSection } from './QRCodeSection';
+export { default as ShopInfoCards } from './ShopInfoCards';
+export { default as PlanFeaturesDisplay } from './PlanFeaturesDisplay';
